inits pageInputWrapper
