+++ src/components/Dashboard/Dashboard.js
@@ -8,7 +8,7 @@ import BoardCard from './BoardCard/BoardCard.js'
 import PageCard from './PageCard/PageCard.js'
 
 import BoardInput from './Input/BoardInput.js'
-import PageInput from './Input/PageInput.js'
+import PageInputWrapper from './Input/PageInputWrapper.js'
 
 import './Dashboard.css'
 
@@ -94,7 +94,7 @@ return (
         
         {/* <BoardInput pushBoard={pushBoard} /> */}
 
-        <PageInput 
+        <PageInputWrapper 
         pushPage={pushPage}
         nestSeq={[]}
         />
