+++ src/components/Dashboard/Input/PageInput.js
@@ -1,83 +1,73 @@
-import './_PageInput.css'
-import React, { useState, useEffect, useRef } from 'react'
+// import './_PageInput.css'
+// import React, { useState, useEffect, useRef } from 'react'
 
-function PageInput({ pushPage, nestSeq }) {
+// function PageInput({ pushPage, nestSeq, pageInputRef }) {
 
-    const [page, setPage] = useState('')
-    const [addingPage, setAddingPage] = useState(false)
-    const [lock, setLock] = useState(false)
+//     const [page, setPage] = useState('')
+//     // const [addingPage, setAddingPage] = useState(false)
+//     // i need a state action that sets addingPageFalse
 
-    const pageInputRef = useRef()
+//     const [lock, setLock] = useState(false)
 
-    useEffect(() => {
-        if(addingPage) {
-            setLock(false)
-            setTimeout(() => {
-                pageInputRef.current.focus()
-            }, .0001)
-        }
-    }, [addingPage])
+//     useEffect(() => {
+//         if(addingPage) {
+//             setLock(false)
+//             setTimeout(() => {
+//                 pageInputRef.current.focus()
+//             }, .0001)
+//         }
+//     }, [addingPage])
 
-    const handleWrite = (ev) => {
-        setPage(ev.target.value)
-    }
+//     const handleWrite = (ev) => {
+//         setPage(ev.target.value)
+//     }
 
-    const handleBlur = () => {
-        if(!lock) {
-            setAddingPage(false)
-        }
-    }
+//     const handleBlur = () => {
+//         if(!lock) {
+//             // setAddingPage(false)
+//             // use state action here
+//         }
+//     }
 
-    const lockUnlock = () => {
-        setLock(true)
-        setTimeout(() => {
-            setLock(false)
-        }, .0001)
-    }
+//     const lockUnlock = () => {
+//         setLock(true)
+//         setTimeout(() => {
+//             setLock(false)
+//         }, .0001)
+//     }
 
-    const handleSave = () => {
+//     const handleSave = () => {
 
-        if (page.length > 0) {
-            pushPage(nestSeq ? nestSeq[nestSeq.length-1] : null, page)
-            setAddingPage(false)
-        }
+//         if (page.length > 0) {
+//             pushPage(nestSeq ? nestSeq[nestSeq.length-1] : null, page)
+//             setAddingPage(false)
+//         }
 
-    }
+//     }
 
-return (
-<>
-{
-    !addingPage
-    ?
-    (
-        <div className='pageInput_notAddingPage'>
-            <p 
-            className='notAddingPage_btn'
-            onClick={() => {setAddingPage(true)}}>+ Add Page</p>
-        </div>
-    )
-    :
-    (
-        <div 
-        className='pageInput_addingPage'
-        onMouseDown={(ev) => ev.preventDefault()}
-        onBlur={handleBlur}
-        onClick={lockUnlock}
-        >
-            <input 
-            className='addingPage_input'
-            onChange={handleWrite}
-            value={page}
-            ref={pageInputRef}
-            />
-            <div 
-            className='addingPage_saveBtn'
-            onClick={handleSave}>+ Save Page</div>
-        </div>
-    )
-}
-</>
-)
-}
+// return (
+// <>
 
-export default PageInput
+//     <div 
+//     className='pageInput_addingPage'
+//     onMouseDown={(ev) => ev.preventDefault()}
+//     onBlur={handleBlur}
+//     onClick={lockUnlock}
+//     >
+//         <input 
+//         className='addingPage_input'
+//         onChange={handleWrite}
+//         value={page}
+//         ref={pageInputRef}
+//         />
+//         <div 
+//         className='addingPage_saveBtn'
+//         onClick={handleSave}>+ Save Page</div>
+//     </div>
+
+
+// </>
+// )
+// }
+
+// export default PageInput
